feat: add client patches
