+++ instrumentation/ruby_kafka/lib/opentelemetry/instrumentation/ruby_kafka/version.rb
@@ -7,7 +7,7 @@
 module OpenTelemetry
   module Instrumentation
     module RubyKafka
-      VERSION = '0.8.0'
+      VERSION = '0.9.0'
     end
   end
 end
