+++ instrumentation/ruby_kafka/test/test_helper.rb
@@ -25,3 +25,15 @@ def clear_notification_subscriptions
     ActiveSupport::Notifications.unsubscribe(event::EVENT_NAME)
   end
 end
+
+def wait_for(max_attempts: 10, retry_delay: 0.10, error_message:)
+  attempts = 0
+  while attempts < max_attempts
+    return if yield
+
+    attempts += 1
+    raise error_message if attempts >= max_attempts
+
+    sleep retry_delay
+  end
+end
